refactor: rename "simple" to "single" to match file pattern under nodes
