+++ src/types/typeGuards/index.ts
@@ -1,3 +1,3 @@
 export * from "./compound";
 export * from "./objects";
-export * from "./simple";
+export * from "./single";
