+++ src/types/typeGuards/single.ts
